requirements.txt
pyserial>=3.5
cantools>=39.0.0
